My Experiment Pipeline - TSWM
Hello :)
Just an update, my experiment is now only eyetracking and behavioural as I changed my project from EEG to behavioural. Still the same concept, just no EEG. 
In this task, you will only be going through 5 trials if you set it to no eyetracker on. You will look at a fixation point, and three coloured squares will pop up around the point momentarily. You must remember the colours in your peripheral vision as you are fixating on the fixation point. An array of many squares will appear on the screen and disappear (a distraction). The fixation point will then move to a new location, and you must make an eye movement to the new fixation point. The fixation point will then change colours to one of the stimuli in the presented memory array (the three presented squares), and using your mouse, you will move that point to the exact location that the colour appeared in around the initial fixation point.